+++ packages/dart_frog_cli/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.8
+
+- feat: update brick bundles to dart 3.0.0
+- feat: Add dart-vm-port option to dev command
+- feat: added default port number for --dart-vm-service-port
+- feat: recommend using --dart-vm-service-port on VM port failure
+
 # 0.3.7
 
 - feat: Update to dart 3.0.0 and vg analysis to 5.0.0